Switch all status items to longs

+++ state.h
@@ -6,12 +6,12 @@ extern "C" {
 #endif
 
 struct doorbell_state {
-  volatile unsigned int auto_pulse = 0;
+  volatile long auto_pulse = 0;
 
-  volatile unsigned long ring_at = 0;
-  volatile unsigned long pulse_until = 0;
+  volatile long ring_at = 0;
+  volatile long pulse_until = 0;
 
-  volatile int standing_brightness = 0;
+  volatile long standing_brightness = 0;
 
   bool net_online = false;
   bool snmp_online = false;
